Rename QueryService to DatabaseQueryService. Add interface QueryService. Add SparqlEndpointQueryService
